Bug Fixes & Structuring
