--- sourse/Application/private/App.cpp
@@ -1,2 +1,0 @@
-
-#include "App.h"
